+++ shuttermint/app/events.go
@@ -76,7 +76,7 @@ func MakePrivkeyGeneratedEvent(batchIndex uint64, privkey *ecdsa.PrivateKey) abc
 
 // MakeBatchConfigEvent creates a 'shutter.batch-config' tendermint event. The given
 // startBatchIndex, threshold and list of keyper addresses are encoded as attributes of the event.
-func MakeBatchConfigEvent(startBatchIndex uint64, threshold uint32, keypers []common.Address) abcitypes.Event {
+func MakeBatchConfigEvent(startBatchIndex uint64, threshold uint64, keypers []common.Address) abcitypes.Event {
 	return abcitypes.Event{
 		Type: "shutter.batch-config",
 		Attributes: []kv.Pair{
